feat(contracts): Support multiple Light Account indices per user
This commit introduces a major feature enhancement, enabling support for multiple Light Account instances per user. Previously, the system assumed a single Light Account per user (hardcoded to index 0). This update allows users to have multiple accounts, indexed from 0 upwards, for different purposes such as varied security levels or distinct organizational roles.

The implementation is designed to be fully backward compatible.

### Key Changes

**1. Contract Updates:**

-   **`LightAccountValidatorV1.sol`**:
    -   The core validation logic, including `potentialLightAccountResolvedOwner()` and `_validateLightAccount()`, now accepts a `lightAccountIndex_` parameter. This index is used to deterministically predict the correct Light Account address for a given owner.
    -   A new internal function, `_extractLightAccountIndex()`, has been added to extract the index from ERC-4337 `paymasterAndData`. For backward compatibility, it defaults to index 0 if the index is not provided in the paymaster data.

-   **Pass-through Contracts (`StrategyV1`, `FreezeVotingAzoriusV1`, `FreezeVotingMultisigV1`):**
    -   Functions that initiate actions on behalf of a user (`vote()`, `castFreezeVote()`) have been updated to accept the `lightAccountIndex_`.
    -   This index is then passed down to `potentialLightAccountResolvedOwner()` to ensure the correct user account is resolved.

-   **`StrategyV1ValidatorV1.sol`**:
    -   The validator has been updated to correctly decode and validate the new `vote` function signature, which now includes the `lightAccountIndex_` parameter.

**2. Test Suite Enhancements:**

-   The entire test suite has been updated to reflect the new function signatures, typically by passing `0n` as the default `lightAccountIndex_`.
-   A new, comprehensive test suite has been added to `LightAccountValidatorV1.test.ts` to rigorously test the new multi-index logic. These tests cover:
    -   Resolving owners with different indices.
    -   Extracting the index from paymaster data, including edge cases.
    -   Validation scenarios with matching and mismatched indices.
    -   Backward compatibility by defaulting to index 0.

This feature provides significant new flexibility for users while ensuring that existing integrations continue to function seamlessly.
